+++ @types/mojave.d.ts
@@ -134,7 +134,7 @@ export module mojaveIntegration
     //region Loader
     export interface LoaderInterface
     {
-        start (message: string | null): void;
+        start (...messages: string[]): void;
         end (): void;
     }
     //endregion
@@ -167,9 +167,9 @@ export module mojaveIntegration
 
         /**
          * undefined|null = silent mode, no loader
-         * string         = use the given string as message
+         * string         = use the given string array as messages
          */
-        loading?: string|null;
+        loading?: string[]|null;
 
         /**
          * Flag whether to automatically handle generic request errors: